add Character class
* Warrior inherits from Character
* Change Warrior types to Character in Fight and Party
